Move run script generation logic into the convert package.
This makes using the functions cleaner.
